+++ src/test/java/janggi/domain/piece/SaTest.java
@@ -1,11 +1,11 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.domain.piece.Cha;
-import janggi.domain.piece.Gung;
 import janggi.domain.piece.Pieces;
+import janggi.domain.piece.Sa;
 import janggi.domain.setting.CampType;
 import janggi.domain.value.JanggiPosition;
 import java.util.List;
@@ -17,7 +17,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class GungTest {
+class SaTest {
 
     static final JanggiPosition STANDARD = new JanggiPosition(4, 8);
 
@@ -26,13 +26,13 @@ class GungTest {
     @MethodSource()
     void test1(JanggiPosition destination) {
         //given
-        Gung gung = Gung.generateInitialGung(CampType.CHO).getFirst();
+        Sa sa = Sa.from(STANDARD);
 
         //when
-        Gung movedGung = gung.move(destination, new Pieces(List.of()), new Pieces(List.of()));
+        Sa movedSa = sa.move(destination, new Pieces(List.of()), new Pieces(List.of()));
 
         //then
-        assertThat(movedGung.getPosition()).isEqualTo(destination);
+        assertThat(movedSa.getPosition()).isEqualTo(destination);
     }
 
     static Stream<Arguments> test1() {
@@ -49,10 +49,10 @@ class GungTest {
     @MethodSource()
     void test2(JanggiPosition destination) {
         //given
-        Gung gung = Gung.generateInitialGung(CampType.CHO).getFirst();
+        Sa sa = Sa.from(STANDARD);
 
         //when & then
-        assertThatThrownBy(() -> gung.move(destination, new Pieces(List.of()), new Pieces(List.of())))
+        assertThatThrownBy(() -> sa.move(destination, new Pieces(List.of()), new Pieces(List.of())))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 이동이 불가능합니다.");
     }
@@ -69,12 +69,12 @@ class GungTest {
     @Test
     void test3() {
         //given
-        Gung gung = Gung.generateInitialGung(CampType.CHO).getFirst();
-        JanggiPosition destination = new JanggiPosition(STANDARD.x() + 1, STANDARD.y());
-        Gung otherPiece = Gung.from(destination);
+        Sa sa = Sa.generateInitialSas(CampType.CHO).getFirst();
+        JanggiPosition destination = new JanggiPosition(3, 8);
+        Sa otherPiece = Sa.from(destination);
 
         //when & then
-        assertThatThrownBy(() -> gung.move(destination, new Pieces(List.of()), new Pieces(List.of(otherPiece))))
+        assertThatThrownBy(() -> sa.move(destination, new Pieces(List.of()), new Pieces(List.of(otherPiece))))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 이동이 불가능합니다.");
     }
@@ -83,13 +83,13 @@ class GungTest {
     @Test
     void test4() {
         //given
-        Gung gung = Gung.from(STANDARD);
-        JanggiPosition destination = new JanggiPosition(5, 8);
+        Sa sa = Sa.from(STANDARD);
+        JanggiPosition destination = new JanggiPosition(4,9);
         Cha enemyCha = Cha.from(destination);
         Pieces enemyPieces = new Pieces(List.of(enemyCha));
 
         //when
-        Gung movedGung = gung.move(destination, enemyPieces, new Pieces(List.of()));
+        Sa movedSa = sa.move(destination, enemyPieces, new Pieces(List.of()));
 
         //then
         Assertions.assertThat(enemyPieces.isNotBlockedBy(destination)).isTrue();
@@ -100,10 +100,10 @@ class GungTest {
     @MethodSource()
     void test5(JanggiPosition destination) {
         //given
-        Gung gung = Gung.from(STANDARD);
+        Sa sa = Sa.from(STANDARD);
 
         //when
-        boolean isMoving = gung.ableToMove(destination, new Pieces(List.of()), new Pieces(List.of()));
+        boolean isMoving = sa.ableToMove(destination, new Pieces(List.of()), new Pieces(List.of()));
 
         //then
         Assertions.assertThat(isMoving).isTrue();
@@ -123,10 +123,10 @@ class GungTest {
     @Test
     void test6() {
         //given
-        Gung gung = Gung.from(new JanggiPosition(3,8));
+        Sa sa = Sa.from(new JanggiPosition(3,8));
         JanggiPosition destination = new JanggiPosition(2,8);
         //when
-        boolean isMoving = gung.ableToMove(destination, new Pieces(List.of()), new Pieces(List.of()));
+        boolean isMoving = sa.ableToMove(destination, new Pieces(List.of()), new Pieces(List.of()));
 
         //then
         Assertions.assertThat(isMoving).isFalse();
